make working, but with reliks and some bugs

+++ sources/asm/ft_dis_args_treat.c
@@ -17,7 +17,7 @@ void		ft_dis_args_treat(t_dis *dis_s, t_elem *elems)
 	unsigned	i;
 
 	i = 0;
-	while (i < elems->op->n_args)
+	while (i < (unsigned int)elems->op->n_args)
 	{
 		ft_dis_arg_treat(dis_s, elems, i);
 		i++;
